+++ components/proto-derive/src/lib.rs
@@ -51,37 +51,39 @@ use syn::{Attribute, NestedMeta};
 /// ```
 ///
 /// Corresponding proto file:
-/// ```text
+/// ```proto
 /// message Wallet {
-///  // Public key of the wallet owner.
-///  exonum.crypto.PublicKey pub_key = 1;
-///  // Current balance.
-///  uint64 balance = 2;
+///     // Public key of the wallet owner.
+///     exonum.crypto.PublicKey pub_key = 1;
+///     // Current balance.
+///     uint64 balance = 2;
 /// }
 /// ```
 ///
 /// This macro can also be applied to enums. In proto files enums are represented
 /// by `oneof` field. You can specify `oneof` field name, default is "message".
+/// Corresponding proto file must contain only this oneof field. Possible enum
+/// variants are zero-field and one-field variants.
 /// ```ignore
 /// #[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, ProtobufConvert)]
 /// #[protobuf_convert(source = "schema::runtime::ConfigChange", oneof_field = "message")]
 /// pub enum ConfigChange {
-///    /// New consensus config.
-///    Consensus(ConsensusConfig),
-///    /// New service instance config.
-///    Service(ServiceConfig),
+///     /// New consensus config.
+///     Consensus(ConsensusConfig),
+///     /// New service instance config.
+///     Service(ServiceConfig),
 /// }
 /// ```
 ///
 /// Corresponding proto file:
-/// ```test
+/// ```proto
 /// message ConfigChange {
-///  oneof message {
-///    // New consensus config.
-///    exonum.Config consensus = 1;
-///    // New service instance config.
-///    ServiceConfig service = 2;
-///  }
+///     oneof message {
+///         // New consensus config.
+///         exonum.Config consensus = 1;
+///         // New service instance config.
+///         ServiceConfig service = 2;
+///     }
 /// }
 /// ```
 ///