+++ examples/cryptocurrency-advanced/backend/src/wallet.rs
@@ -15,7 +15,7 @@
 //! Cryptocurrency wallet.
 
 use exonum::crypto::{Hash, PublicKey};
-use exonum_proto_derive::ProtobufConvert;
+use exonum_proto::ProtobufConvert;
 
 use super::proto;