+++ exonum/src/api/node/public/system.rs
@@ -16,11 +16,12 @@
 
 use exonum_merkledb::IndexAccess;
 
+use crate::runtime::ProtoSourceFile;
 use crate::{
     api::{self, node::SharedNodeState, ApiContext, ApiScope},
     blockchain::Schema,
     helpers::user_agent,
-    proto::schema::PROTO_SOURCES as EXONUM_PROTO_SOURCES,
+    proto::schema::{INCLUDES as EXONUM_INCLUDES, PROTO_SOURCES as EXONUM_PROTO_SOURCES},
     runtime::{dispatcher, ArtifactId, InstanceSpec},
 };
 
@@ -139,23 +140,27 @@ impl SystemApi {
 
     fn handle_proto_source(self, name: &'static str, api_scope: &mut ApiScope) -> Self {
         let node_state = self.node_state.clone();
+        let exonum_sources = get_exonum_sources();
         api_scope.endpoint(name, {
             move |query: ProtoSourcesQuery| {
                 if let Some(artifact_id) = query.artifact {
-                    node_state
+                    let sources = node_state
                         .artifact_sources(&artifact_id.parse()?)
                         .ok_or_else(|| {
                             api::Error::NotFound(format!(
                                 "Unable to find sources for artifact {}",
                                 artifact_id
                             ))
-                        })
+                        })?;
+
+                    let mut proto = sources.sources.clone();
+                    proto.extend(filter_exonum_sources(
+                        sources.includes,
+                        exonum_sources.clone(),
+                    ));
+                    Ok(proto)
                 } else {
-                    Ok(EXONUM_PROTO_SOURCES
-                        .as_ref()
-                        .iter()
-                        .map(From::from)
-                        .collect::<Vec<_>>())
+                    Ok(exonum_sources.clone())
                 }
             }
         });
@@ -191,4 +196,30 @@ impl SystemApi {
             .handle_proto_source("v1/proto-sources", api_scope);
         api_scope
     }
+}
+
+fn get_exonum_sources() -> Vec<ProtoSourceFile> {
+    let proto = EXONUM_PROTO_SOURCES
+        .as_ref()
+        .iter()
+        .map(From::from)
+        .collect::<Vec<_>>();
+    let includes = EXONUM_INCLUDES
+        .as_ref()
+        .iter()
+        .map(From::from)
+        .collect::<Vec<_>>();
+
+    proto.into_iter().chain(includes).collect()
+}
+
+fn filter_exonum_sources(
+    files: Vec<ProtoSourceFile>,
+    exonum_sources: Vec<ProtoSourceFile>,
+) -> Vec<ProtoSourceFile> {
+    files
+        .iter()
+        .cloned()
+        .filter(|file| !exonum_sources.contains(file))
+        .collect()
 }