+++ exonum/src/lib.rs
@@ -44,8 +44,6 @@
 extern crate pretty_assertions;
 #[macro_use]
 extern crate exonum_derive;
-#[macro_use]
-extern crate exonum_proto_derive;
 pub extern crate exonum_merkledb;
 #[cfg(feature = "sodiumoxide-crypto")]
 extern crate exonum_sodiumoxide as sodiumoxide;