+++ exonum/src/runtime/mod.rs
@@ -297,8 +297,12 @@ pub trait Runtime: Send + Debug + 'static {
 
     /// Notify the runtime that it has to shutdown.
     ///
-    /// This callback is invoked before the node shutdown, so runtimes can stop themselves
-    /// gracefully.
+    /// This callback is invoked sequentially for each runtime just before the node shutdown,
+    /// so runtimes can stop themselves gracefully.
+    ///
+    /// Invoking of this callback is guaranteed to be the last operation for the runtime.
+    /// Since this method is a part of shutdown process, runtimes can perform blocking and
+    /// heavy operations here if needed.
     fn shutdown(&self) {}
 }
 
@@ -332,18 +336,22 @@ impl From<&(&str, &str)> for ProtoSourceFile {
 /// Artifact Protobuf specification for the Exonum clients.
 #[derive(Debug, Default, Clone, PartialEq)]
 pub struct ArtifactProtobufSpec {
-    /// List of Protobuf files that make up the service interface. The first element in the tuple
-    /// is the file name, the second one is its content.
+    /// List of Protobuf files that make up the service interface.
     ///
     /// The common interface entry point is always in the `service.proto` file.
     pub sources: Vec<ProtoSourceFile>,
+    /// List of service's proto include files.
+    pub includes: Vec<ProtoSourceFile>,
 }
 
-impl From<&[(&str, &str)]> for ArtifactProtobufSpec {
-    fn from(sources_strings: &[(&str, &str)]) -> Self {
-        let sources = sources_strings.iter().map(From::from).collect();
+type ProtoSources<'a> = &'a [(&'a str, &'a str)];
 
-        Self { sources }
+impl<'a> From<(ProtoSources<'a>, ProtoSources<'a>)> for ArtifactProtobufSpec {
+    fn from(sources_strings: (ProtoSources<'a>, ProtoSources<'a>)) -> Self {
+        let sources = sources_strings.0.iter().map(From::from).collect();
+        let includes = sources_strings.1.iter().map(From::from).collect();
+
+        Self { sources, includes }
     }
 }