+++ exonum/tests/websocket/blockchain/mod.rs
@@ -25,7 +25,7 @@ use exonum::{
     },
 };
 use exonum_merkledb::{Snapshot, TemporaryDB};
-use exonum_proto_derive::ProtobufConvert;
+use exonum_proto::ProtobufConvert;
 
 use std::{
     net::SocketAddr,